refactor to remove matchfound and replace with personcompleted event and text fix
