+++ app/DynamicsAdapter/DynamicsAdapter.Web/PersonSearch/Models/PersonAcceptedEvent.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace DynamicsAdapter.Web.PersonSearch.Models
+{
+    public class PersonAcceptedEvent
+    {
+        public Guid SearchRequestId { get; set; }
+        public DateTime TimeStamp { get; set; }
+        public ProviderProfile ProviderProfile { get; set; }
+    }
+}
