+++ app/DynamicsAdapter/DynamicsAdapter.Web/PersonSearch/Models/PersonCompletedEvent.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace DynamicsAdapter.Web.PersonSearch.Models
+{
+    public class PersonCompletedEvent
+    {
+        public Guid SearchRequestId { get; set; }
+        public DateTime TimeStamp { get; set; }
+        public ProviderProfile ProviderProfile { get; set; }
+        public Person Person { get; set; }
+
+        public IEnumerable<PersonId> PersonIds { get; set; }
+    }
+}
