+++ app/DynamicsAdapter/DynamicsAdapter.Web/PersonSearch/Models/ProviderProfile.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace DynamicsAdapter.Web.PersonSearch.Models
+{
+    public class ProviderProfile
+    {
+        public string Name { get; set; }
+    }
+}
